Rewrite container to be nix-based
The old Dockerfile no longer works as the image postgres:12 is based off
doesn't have access to the multicorn debian packages anymore.  Since we
already have a fully functional nix-based environment for development,
it makes sense to just roll that into the docker container... but we are
losing all the advanced env-variable based config that the old docker
container did.  It's probably fine given the scope and usage of this
tool.